#703 More refactoring, added uml tests, added JAssert dependency
