+++ cobigen-templates/templates-oasp4j/crud_java_ea_uml/templates/java/${variables.rootPackage}/${variables.component}/common/api/${variables.className}.java.ftl
@@ -16,7 +16,7 @@ import java.util.List;
 
 public interface ${name} extends ApplicationEntity {
 
-${UmlUtil.generateConnectorsVariablesMethodsText(false,false,variables.className)}
+${UmlUtil.generateConnectorsVariablesMethodsText(false,false)}
 
 <#list elemDoc["./ownedAttribute"] as attribute>
   <#if (attribute["@name"])??>
