+++ cobigen-templates/templates-oasp4j/pom.xml
@@ -44,5 +44,12 @@
       <artifactId>openapiplugin-model</artifactId>
       <version>2.0.0</version>
     </dependency>
+    <dependency>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+      <version>3.10.0</version>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 </project>
