--- cobigen-templates/templates-oasp4j/src/main/java/com/devonfw/cobigen/templates/oasp4j/utils/uml/Connectors.java
@@ -1,213 +1,0 @@
-package com.devonfw.cobigen.templates.oasp4j.utils.uml;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.apache.commons.lang3.text.WordUtils;
-
-import com.devonfw.cobigen.templates.oasp4j.utils.OaspUtil;
-
-/**
- * Class that contains every connector found for one class and generates the resultant text for the template.
- */
-public class Connectors {
-
-    private List<Connector> connectors;
-
-    /**
-     * The constructor.
-     */
-    public Connectors() {
-
-        connectors = new ArrayList<Connector>();
-    }
-
-    /**
-     * @param connector
-     *            class storing connected information
-     */
-    public void addConnector(Connector connector) {
-
-        connectors.add(connector);
-    }
-
-    /**
-     * @param isImpl
-     *            true if we are the Implementation
-     * @param isOverride
-     *            true if we are Entity
-     * @param className
-     *            name of the class
-     * @return Generated text
-     */
-    public String generateText(boolean isImpl, boolean isOverride, String className) {
-
-        String content = "";
-        if (isImpl) {
-            for (Connector connector : connectors) {
-                String connectedClassName = connector.getCounterpartName();
-                String multiplicity = connector.getCounterpartMultiplicity();
-                if (multiplicity == null || multiplicity.equals("1")) {
-                    content +=
-                        "\n\n\tprivate " + connectedClassName + "Entity " + connectedClassName.toLowerCase() + ";";
-                } else if (multiplicity != null && multiplicity.equals("*")) {
-                    content += "\n\n\tprivate List<" + connectedClassName + "Entity> "
-                        + new OaspUtil().removePlural(connectedClassName.toLowerCase()) + "s;";
-                }
-            }
-        }
-
-        for (Connector connector : connectors) {
-            String connectedClassName = connector.getCounterpartName();
-            String multiplicity = connector.getCounterpartMultiplicity();
-            if (multiplicity == null || multiplicity.equals("1")) {
-
-                content += "\n\n\t";
-                if (isOverride) {
-                    content += "@Override\n\t";
-                }
-                if (isImpl) {
-                    content += getRelationshipAnnotations(connector) + "\n\t";
-                    content += "public " + connectedClassName + "Entity get" + connectedClassName + "()";
-                } else {
-                    content += "public Long get" + connectedClassName + "Id()";
-                }
-                if (isImpl) {
-                    content += "{" + "\n\t\treturn this." + connectedClassName.toLowerCase() + ";" + "\n\t}";
-                } else {
-                    content += ";";
-                }
-
-                content += "\n\n\t";
-                if (isOverride) {
-                    content += "@Override\n\t";
-                }
-                if (isImpl) {
-                    content += "public void set" + connectedClassName + "(" + connectedClassName + "Entity "
-                        + connectedClassName.toLowerCase() + ")";
-                } else {
-                    content +=
-                        "public void set" + connectedClassName + "Id(Long " + connectedClassName.toLowerCase() + "Id)";
-                }
-                if (isImpl) {
-                    content += "{" + "\n\t\tthis." + connectedClassName.toLowerCase() + " = "
-                        + connectedClassName.toLowerCase() + ";" + "\n\t}";
-                } else {
-                    content += ";";
-                }
-                // Now w generate the get and set IDs methods for the implementation
-                if (isImpl) {
-                    // getter
-                    content += "\n\n\t";
-                    content += "@Override\n\t";
-                    content += "public Long get" + connectedClassName + "Id()";
-                    content += "{" + "\n\t\tif(this." + connectedClassName.toLowerCase() + " == null){";
-                    content += "\n\t\treturn null;\n\t}";
-                    content += "\n\t\treturn this." + connectedClassName.toLowerCase() + ".getId();" + "\n\t}";
-                    // setter
-                    content += "\n\n\t";
-                    content += "@Override\n\t";
-                    content +=
-                        "public void set" + connectedClassName + "Id(Long " + connectedClassName.toLowerCase() + "Id)";
-                    content += "{" + "\n\t\tif(" + connectedClassName.toLowerCase() + "Id == null){";
-                    content += "\n\t\tthis." + connectedClassName.toLowerCase() + " = null;\n\t}";
-                    content += "else {\n\t";
-                    content += connectedClassName + "Entity " + connectedClassName.toLowerCase() + "Entity = new "
-                        + connectedClassName + "Entity();\n\n\t";
-                    content +=
-                        connectedClassName.toLowerCase() + ".setId(" + connectedClassName.toLowerCase() + "Id);\n\n\t";
-                    content += "this." + connectedClassName.toLowerCase() + " " + "= "
-                        + connectedClassName.toLowerCase() + "Entity;\n\n\t}";
-                    content += "\n\n\t}";
-                }
-
-            } else if (multiplicity != null && multiplicity.equals("*") && isImpl) {
-
-                content += "\n\n\t";
-                if (isOverride) {
-                    content += "@Override\n\t";
-                }
-                content += getRelationshipAnnotations(connector) + "\n\t";
-                content += "public List<" + connectedClassName + "Entity> get"
-                    + new OaspUtil().removePlural(connectedClassName) + "s()";
-                content += "{" + "\n\t\treturn this." + new OaspUtil().removePlural(connectedClassName.toLowerCase())
-                    + "s;" + "\n\t}";
-                content += "\n\n\t";
-                if (isOverride) {
-                    content += "@Override\n\t";
-                }
-                content +=
-                    "public void set" + new OaspUtil().removePlural(connectedClassName) + "s(List<" + connectedClassName
-                        + "Entity> " + new OaspUtil().removePlural(connectedClassName.toLowerCase()) + "s)";
-                content += "{" + "\n\t\tthis." + new OaspUtil().removePlural(connectedClassName.toLowerCase()) + "s = "
-                    + new OaspUtil().removePlural(connectedClassName.toLowerCase()) + "s;" + "\n\t}";
-            }
-        }
-        return content;
-    }
-
-    /**
-     * Generates the annotations of all the connected classes
-     *
-     * @return relationship string with all the annotations for the connected classes
-     */
-    private String getRelationshipAnnotations(Connector source) {
-
-        String relationship = "";
-        if (source.ISSOURCE) {
-            if (source.getMultiplicity() == null || source.getMultiplicity().equals("1")) {
-                if (source.getCounterpartMultiplicity() == null || source.getCounterpartMultiplicity().equals("1")) {
-                    relationship = "@OneToOne()" + "\n\t@JoinColumn(name = \"" + source.getCounterpartName() + "Id\")";
-                } else if (source.getCounterpartMultiplicity().equals("*")) {
-                    relationship = "@OneToMany(fetch = FetchType.LAZY)\n\t@JoinColumn(name = \""
-                        + WordUtils.capitalize(source.getCounterpartName()) + "id\")";
-                }
-            } else if (source.getMultiplicity().equals("*")) {
-                if (source.getCounterpartMultiplicity().equals("*")) {
-                    relationship += "@ManyToMany()";
-                    relationship += "\n\t@JoinTable(name = \"" + WordUtils.capitalize(source.getCounterpartName())
-                        + WordUtils.capitalize(source.getClassName()) + "\", joinColumns = @JoinColumn(name = \""
-                        + source.getClassName() + "Id\"), inverseJoinColumns = @JoinColumn(name = \""
-                        + source.getCounterpartName() + "Id\"))";
-                } else if (source.getCounterpartMultiplicity().equals("1")) {
-                    relationship += "@ManyToOne(fetch = FetchType.LAZY)\n\t@JoinColumn(name = \""
-                        + source.getCounterpartName() + "Id\")";
-                }
-            }
-        } else if (source.ISTARGET) {
-            if (source.getCounterpartMultiplicity() == null || source.getCounterpartMultiplicity().equals("1")) {
-                if (source.getMultiplicity() == null || source.getMultiplicity().equals("1")) {
-                    relationship = "@OneToOne()" + "\n\t@JoinColumn(name = \"" + source.getCounterpartName() + "Id\")";
-                } else if (source.getMultiplicity().equals("*")) {
-                    relationship += "@ManyToOne(fetch = FetchType.LAZY)\n\t@JoinColumn(name = \""
-                        + source.getCounterpartName() + "Id\")";
-                }
-            } else if (source.getCounterpartMultiplicity().equals("*")) {
-                if (source.getMultiplicity().equals("*")) {
-                    relationship += "@ManyToMany(mappedBy = \""
-                        + new OaspUtil().removePlural(source.getClassName()).toLowerCase() + "s\")";
-                } else if (source.getMultiplicity().equals("1")) {
-                    relationship = "@OneToMany(fetch = FetchType.LAZY, mappedBy = \""
-                        + source.getCounterpartName().toLowerCase() + "\")";
-                }
-            }
-        }
-        return relationship;
-    }
-
-    /**
-     * Stores all the connected classes
-     *
-     * @return connectedClasses array with connected classes
-     */
-    public ArrayList<String> getConnectedClasses() {
-
-        ArrayList<String> connectedClasses = new ArrayList<String>();
-
-        for (Connector connector : connectors) {
-            connectedClasses.add(connector.getClassName());
-        }
-
-        return connectedClasses;
-    }
-}
